backend/requirements.txt
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20200713
Django==3.2
django-allauth==0.45.0
django-cleanup==6.0.0
django-cors-headers==3.8.0
django-cpf-cnpj==1.0.0
django-environ==0.6.0
django-extensions==3.1.3
django-filter==21.1
django-import-export==2.7.1
django-multitenant==4.1.1
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-simple-history==3.3.0
django-templated-mail==1.1.1
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
jsonschema-to-openapi==0.2.1
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
redis==4.1.4
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.1.0
typing_extensions==4.5.0
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
